disney: Add (slightly hacky) support for season listings
